Hides HadoopFsRelation related data source API
